UserPage 페이지 추가

+++ backend/src/main/java/com/ll/cafeservice/domain/order/dto/response/OrderItemResponse.java
@@ -4,5 +4,5 @@ public record OrderItemResponse(
         Long id,
         String name,
         int quantity,
-        Long price
+        Integer price
 ) { }
